Add type alias to vary numerical precision
5-10% performance gain by using f32 instead of f64. Loss in precision does not significantly affect quality of results.

When f16 is natively supported in rust[0], we should experiment with an even lower precision.

0. https://internals.rust-lang.org/t/adding-minifloats-f24-f16-f8-as-native-types/2367/12

+++ src/network.rs
@@ -4,6 +4,8 @@ extern crate vec_map;
 use self::vec_map::VecMap;
 use self::vec_map::Entry::{Vacant, Occupied};
 
+use Float;
+
 use neuron::Neuron;
 use synapse::Synapse;
 use spike::Spike;
@@ -25,7 +27,7 @@ pub struct Network<N: Neuron, S: Synapse> {
     next_neuron_id: usize,
     next_synapse_id: usize,
 
-    now: f64,
+    now: Float,
 }
 
 impl<N: Neuron, S: Synapse> Network<N, S> {
@@ -86,13 +88,13 @@ impl<N: Neuron, S: Synapse> Network<N, S> {
         Ok(synapse_id)
     }
 
-    pub fn dump_weights(&self, weights: &mut [f64]) {
+    pub fn dump_weights(&self, weights: &mut [Float]) {
         for (i, s) in self.synapses.iter() {
             weights[i] = s.weight();
         }
     }
 
-    pub fn tick(&mut self, ticks: usize, inputs: &[f64], outputs: &mut [f64]) -> f64 {
+    pub fn tick(&mut self, ticks: usize, inputs: &[Float], outputs: &mut [Float]) -> Float {
         // let mut post_recv_count = 0;
         // let mut pre_recv_count = 0;
         let neuron_count = self.neurons.len();
