+++ src/sym/synapse.rs
@@ -1,22 +1,23 @@
+use Float;
 use synapse::Synapse;
 use sym::config::SymConfig;
 
 #[derive(Debug, Clone, Copy)]
 pub struct SymSynapse {
-  weight: f64,
+  weight: Float,
 
-  min: f64,
-  max: f64,
+  min: Float,
+  max: Float,
 
-  a_sym: f64,
+  a_sym: Float,
 
-  tau_a: f64,
-  tau_b: f64,
+  tau_a: Float,
+  tau_b: Float,
 
   delay: usize,
 
-  pre_time: f64,
-  post_time: f64,
+  pre_time: Float,
+  post_time: Float,
 }
 
 impl SymSynapse {
@@ -34,12 +35,12 @@ impl SymSynapse {
     }
   }
 
-  fn get_delta(&self) -> f64 {
+  fn get_delta(&self) -> Float {
     let dt = self.post_time - self.pre_time;
     self.a_sym * (1.0 - (dt / self.tau_a).powi(2)) * (-dt.abs() / self.tau_b).exp()
   }
 
-  fn integrate(&mut self, delta: f64) {
+  fn integrate(&mut self, delta: Float) {
     self.weight = self.weight + delta;
     if self.weight > self.max {
       self.weight = self.max;
@@ -50,7 +51,7 @@ impl SymSynapse {
 }
 
 impl Synapse for SymSynapse {
-  fn weight(&self) -> f64 {
+  fn weight(&self) -> Float {
     self.weight
   }
 
@@ -58,7 +59,7 @@ impl Synapse for SymSynapse {
     self.delay
   }
 
-  fn pre_recv(&mut self, now: f64) -> f64 { // delta
+  fn pre_recv(&mut self, now: Float) -> Float { // delta
       self.pre_time = now;
 
       let delta = self.get_delta();
@@ -66,7 +67,7 @@ impl Synapse for SymSynapse {
       delta
   }
 
-  fn post_recv(&mut self, now: f64) -> f64 { // delta
+  fn post_recv(&mut self, now: Float) -> Float { // delta
       self.post_time = now;
 
       let delta = self.get_delta();
